Peer list support
Back to the functionnality of the old version. In less files.
Uses the nice qt view model system.
A list of peers is updated by discovery::browser.
This lists supports dropping urls (starts upload to the peer).
You can also select peers and click "send file", which lets you select a
file and send it to the group.