Support of restartable discovery.
Discovery objects can now be started independently of the rest.
They can be restarted on failure.
The statusbar not shows a button to restart them if they fail, and show
any error from them.
Discovered peers are now inreactive QObject that might be updated and
are destroyed if the browser fails.