Payload to chunk system.
Allow to encode a file hierarchy as a sequence of chunks, and the
opposite.
It is still very rigid: chunks must be in order, no restart, etc.
But at least it should fix the performance problem and enable directory
transfer.

Note that it will remove one functionnality: the ability to select the
exact output filename for the 1-file case.
We can only select the directory where the file will be put.